fixed Makefile maybe

+++ libft/srcs/ft_gridset.c
@@ -6,15 +6,15 @@
 /*   By: svan-der <svan-der@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/05/16 12:41:41 by svan-der       #+#    #+#                */
-/*   Updated: 2019/10/15 13:30:01 by svan-der      ########   odam.nl         */
+/*   Updated: 2019/11/12 17:20:31 by svan-der      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "includes/libft.h"
+#include "libft.h"
 
 void		ft_gridset(void **grid, char c, size_t y, size_t x)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < y)
